+++ src/PAL/COM/sockets/ssl/openssl/ssl_exit_context_internal.cpp
@@ -15,7 +15,7 @@
 
 bool ssl_exit_context_internal(int sslContextHandle )
 {
-    if((sslContextHandle >= ARRAYSIZE(g_SSL_Driver.m_sslContextArray)) || (sslContextHandle < 0) || (g_SSL_Driver.m_sslContextArray[sslContextHandle].SslContext == NULL))
+    if((sslContextHandle >= (int)ARRAYSIZE(g_SSL_Driver.m_sslContextArray)) || (sslContextHandle < 0) || (g_SSL_Driver.m_sslContextArray[sslContextHandle].SslContext == NULL))
     {
         return FALSE;
     }
